update
moved the markup to another file and took a global appeal to local storage
